BREAKING CHANGE use ext marks instead of signs
All highlights are now done with ext marks.

The setup api has following breaking changes.

`headline_signs` is renamed to `headline_highlights` and takes a list of
highlight groups instead of a list of signs.

`codeblock_sign` is renamed to `codeblock_highlight` and takes a
highlight group instead of a sign.