fix: ensure fat headlines are not shifted by conflicting extmarks
This commit makes `headlines.nvim` compatible with Nvim Orgmode's new
Virtual Indent option added by
https://github.com/nvim-orgmode/orgmode/pull/627.

This also will ensure any features like it in the future from other
plugins do not interfere with the headlines later on.

In theory, extmark priorities would work here, but in pratice they do
not. If this extmark is set *last* by a higher priority using only the
virtual text, it can be arbitrarily shifted by other extmark content and
the same occurs when the extmark set is of a lower priority. By
specifying the window column to apply to, we bypass all the priorities
shenanigans and directly set the correct position of the headline.